refactor the copy function of EditableQuestionaireResponseForm
